Widen length to 64 bits before shifting. Corrects SHA1 sums computed on blocks between 512MB and 4GB in size. Blocks of 4GB and larger are not supported yet.
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12753 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/utils/sha1.cc
@@ -252,7 +252,7 @@ void SHA1::update(const byte* data, unsigned len)
 	assert(!m_finalized);
 	uint32 j = (m_count >> 3) & 63;
 
-	m_count += len << 3;
+	m_count += uint64(len) << 3;
 
 	uint32 i;
 	if ((j + len) > 63) {
